Updated the code and made them more similar, and fixed a bug

+++ ManagedANGLE/Gles.cs
@@ -7,9 +7,9 @@ using GLboolean = System.Byte;
 using GLenum = System.UInt32;
 using GLfloat = System.Single;
 using GLint = System.Int32;
-using GLintptr = System.Int64;
+using GLintptr = System.Int32;
 using GLsizei = System.Int32;
-using GLsizeiptr = System.Int64;
+using GLsizeiptr = System.Int32;
 using GLuint = System.UInt32;
 
 namespace ManagedANGLE
@@ -76,7 +76,6 @@ namespace ManagedANGLE
             if (error != 0)
                 throw new Exception("OpenGL ES error: 0x" + error.ToString("x").PadLeft(4, '0'));
         }
-
 
         [DllImport(libGLESv2)]
         public static extern GLuint glCreateProgram();
